requirements.txt
aniso8601==6.0.0
argon2-cffi==19.1.0
cffi==1.12.3
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
pytz==2019.1
six==1.12.0
Werkzeug==0.15.2
